# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/match/accept_invite": {
            "post": {
                "description": "accept an invitation based on the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match service"
                ],
                "summary": "accept an invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "description": "Accept Invite Request Body",
                        "name": "AcceptInviteReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apihttp.AcceptInviteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concrete Error Msg",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/match/cancel_match/{uid}": {
            "post": {
                "description": "cancel match",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match service"
                ],
                "summary": "cancel match",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "player uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concrete Error Msg",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/match/change_role": {
            "post": {
                "description": "change a player's role based on the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match service"
                ],
                "summary": "change a player's role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "description": "Change Role Request Body",
                        "name": "ChangeRoleReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apihttp.ChangeRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concrete Error Msg",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/match/create_group": {
            "post": {
                "description": "create a new group based on the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match service"
                ],
                "summary": "create a new group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "description": "Create Group Request Body",
                        "name": "CreateGroup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pto.CreateGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/match/dissolve_group/{uid}": {
            "post": {
                "description": "dissolve a group based on the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match service"
                ],
                "summary": "dissolve a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concrete Error Msg",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/match/enter_group": {
            "post": {
                "description": "enter a group based on the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match service"
                ],
                "summary": "enter a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "description": "Enter Group Request Body",
                        "name": "EnterGroupReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apihttp.EnterGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concrete Error Msg",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/match/exit_group/{uid}": {
            "post": {
                "description": "exit a group based on the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match service"
                ],
                "summary": "exit a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concrete Error Msg",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/match/invite": {
            "post": {
                "description": "invite a player based on the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match service"
                ],
                "summary": "invite a player",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "description": "Invite Request Body",
                        "name": "InviteReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apihttp.InviteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concrete Error Msg",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/match/kick_player": {
            "post": {
                "description": "kick a player based on the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match service"
                ],
                "summary": "kick a player",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "description": "Kick Player Request Body",
                        "name": "KickPlayerReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apihttp.KickPlayerReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concrete Error Msg",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/match/refuse_invite": {
            "post": {
                "description": "refuse an invitation based on the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match service"
                ],
                "summary": "refuse an invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "description": "Refuse Invite Request Body",
                        "name": "RefuseInviteReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apihttp.RefuseInviteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concrete Error Msg",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/match/set_nearby_join_group": {
            "post": {
                "description": "set whether group can be entered from nearby players list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match service"
                ],
                "summary": "set nearby join group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "description": "Set Nearby Join Group Request Body",
                        "name": "SetNearbyJoinGroupReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apihttp.SetNearbyJoinGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concrete Error Msg",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/match/set_recent_join_group": {
            "post": {
                "description": "set whether group can be entered from recent players list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match service"
                ],
                "summary": "set recent join group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "x-request-id",
                        "in": "header"
                    },
                    {
                        "description": "Set Recent Join Group Request Body",
                        "name": "SetRecentJoinGroupReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apihttp.SetRecentJoinGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concrete Error Msg",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apihttp.AcceptInviteReq": {
            "type": "object",
            "required": [
                "group_id",
                "invitee_info",
                "inviter_uid"
            ],
            "properties": {
                "group_id": {
                    "type": "integer"
                },
                "invitee_info": {
                    "$ref": "#/definitions/pto.PlayerInfo"
                },
                "inviter_uid": {
                    "type": "string"
                }
            }
        },
        "apihttp.ChangeRoleReq": {
            "type": "object",
            "required": [
                "captain_uid",
                "role",
                "target_uid"
            ],
            "properties": {
                "captain_uid": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/entry.GroupRole"
                },
                "target_uid": {
                    "type": "string"
                }
            }
        },
...
        "pto.CreateGroup": {
            "type": "object",
            "required": [
                "game_mode",
                "mode_version",
                "uid"
            ],
            "properties": {
                "game_mode": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/constant.GameMode"
                        }
                    ],
                    "example": 905
                },
                "glicko2_info": {
                    "$ref": "#/definitions/pto.Glicko2Info"
                },
                "mode_version": {
                    "type": "integer",
                    "example": 1
                },
                "rank": {
                    "type": "integer"
                },
                "star": {
                    "type": "integer"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
...
    }
}